% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deployrInput.R
\name{revoInput}
\alias{revoInput}
\title{Define R Script Inputs \bold{deprecated}.}
\usage{
revoInput(content)
}
\arguments{
\item{content}{A valid JSON string configuration.}
}
\description{
This function is now \bold{deprecated}. Please transition to 
\code{\link{deployrInput}}.
}

