% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deployrPackage.R
\name{revoPackage}
\alias{revoPackage}
\title{Install and load R packages. \bold{deprecated}.}
\usage{
revoPackage(pkgs, repos = getOption("repos"), ...)
}
\arguments{
\item{pkgs}{package nam}

\item{repos}{test code containing expectations}

\item{...}{additional arguments to pass to \code{install.packages()}.}
}
\description{
This function is now \bold{deprecated}. Please transition to 
\code{\link{deployrPackage}}.
}

